refactor(app): move Breadcrumbs
